organize files a bit and rename a few of the classes
