+++ src/seagulls/examples/__init__.py
@@ -1,16 +1,12 @@
-from ._main_menu_scene import (
-    MainMenuBackground,
-    MainMenuScene,
-    MainMenuSceneManager,
-    WindowScene
-)
-from ._session import (AsyncGameSession, BlockingGameSession)
+from seagulls.examples._scene_manager import ExampleSceneManager
+from seagulls.examples._simple_stars_background import SimpleStarsBackground
+from seagulls.examples._window_scene import WindowScene
+
+from ._main_menu_scene import MainMenuScene
+from ._session import AsyncGameSession, BlockingGameSession
 
 __all__ = [
-    "MainMenuBackground",
     "MainMenuScene",
-    "MainMenuSceneManager",
     "AsyncGameSession",
-    "BlockingGameSession",
-    "WindowScene"
+    "BlockingGameSession"
 ]
