+++ src/seagulls/examples/seagulls/_scene.py
